feat: rename "addons" folder into "hooks" (#54)
To fix a bug with Tailwind 4 not being able to resolve
a child folder named the same as a parent folder within
a glob expression, this change is needed.

To also better communicate which files are used in which hook,
the subsequent folders are renamed to match the (upper camel)
hook name.